+++ pandas/tests/reshape/test_merge.py
@@ -592,9 +592,9 @@ class TestMerge(object):
                        index=index, columns=['pnum'])
         df2 = concat([df, df])
         result = df.merge(df2, left_index=True, right_index=True, how='inner')
-        expected = DataFrame(np.tile(np.arange(16).repeat(2).reshape(-1, 1), 2),
-                             columns=['pnum_x', 'pnum_y'],
-                             index=df2.sort_index().index)
+        expected = DataFrame(
+            np.tile(np.arange(16, dtype=np.int64).repeat(2).reshape(-1, 1), 2),
+            columns=['pnum_x', 'pnum_y'], index=df2.sort_index().index)
         tm.assert_frame_equal(result, expected)
 
     def test_merge_on_periods(self):
